+++ src/test/java/nextstep/security/authentication/provider/UsernamePasswordAuthenticationProviderTest.java
@@ -5,9 +5,9 @@ import nextstep.security.authentication.token.UsernamePasswordAuthenticationToke
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static nextstep.security.MockFactory.PASSWORD;
-import static nextstep.security.MockFactory.USERNAME;
-import static nextstep.security.MockFactory.createUsernamePasswordAuthenticationProvider;
+import static nextstep.security.Fixture.PASSWORD;
+import static nextstep.security.Fixture.USERNAME;
+import static nextstep.security.Fixture.createUsernamePasswordAuthenticationProvider;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
