+++ src/test/java/nextstep/security/authentication/token/UsernamePasswordAuthenticationTokenTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static nextstep.security.MockFactory.PASSWORD;
-import static nextstep.security.MockFactory.USERNAME;
+import static nextstep.security.Fixture.PASSWORD;
+import static nextstep.security.Fixture.USERNAME;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class UsernamePasswordAuthenticationTokenTest {
