+++ src/test/java/nextstep/security/authentication/token/basic/BasicAuthenticationTokenTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Base64;
 
-import static nextstep.security.MockFactory.PASSWORD;
-import static nextstep.security.MockFactory.USERNAME;
+import static nextstep.security.Fixture.PASSWORD;
+import static nextstep.security.Fixture.USERNAME;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
