+++ src/test/java/nextstep/security/filter/BasicAuthorizationFilterTest.java
@@ -11,8 +11,8 @@ import org.springframework.mock.web.MockHttpServletResponse;
 
 import java.io.IOException;
 
-import static nextstep.security.MockFactory.createBasicRequest;
-import static nextstep.security.MockFactory.createUserDetailsService;
+import static nextstep.security.Fixture.createBasicRequest;
+import static nextstep.security.Fixture.createUserDetailsService;
 import static nextstep.security.Steps.testAuthentication;
 
 class BasicAuthorizationFilterTest {
